🐛 fix dropdown issue on multiselect
The dropdown bug which was discovered on the multiselect component while deselecting all items but one would close the dropdown and bring it back again while deselecting / removing the last item. This only happened when deselecting through clicking on chip's close icon. This was happening due to the click event propagating to the chip's parent which also responds to the click event and toggles the menu. The issue is fixed now by stopping the event from bubbling up to the parent element. #135